requirements.txt
numpy
pandas
scikit-learn
xgboost
pyarrow
zstandard
matplotlib
plotly
pymongo
python-dotenv
